use tmp instead of uploads

+++ server.js
@@ -7,12 +7,8 @@ const path = require('path');
 
 const app = express();
 
-const uploadDir = 'uploads';
-if (!fs.existsSync(uploadDir)) {
-    fs.mkdirSync(uploadDir);
-}
-
-const upload = multer({ dest: uploadDir });
+// USAR /tmp EN RENDER
+const upload = multer({ dest: '/tmp' });
 
 ffmpeg.setFfmpegPath(ffmpegPath);
 
@@ -20,19 +16,24 @@ ffmpeg.setFfmpegPath(ffmpegPath);
 app.use(express.static(__dirname));
 
 app.post('/convert', upload.single('audio'), (req, res) => {
+
+    if (!req.file) {
+        return res.status(400).send('No se recibió archivo');
+    }
+
     const inputPath = req.file.path;
     const outputPath = `${inputPath}.mp3`;
 
     ffmpeg(inputPath)
         .toFormat('mp3')
         .on('end', () => {
-            res.download(outputPath, () => {
+            res.download(outputPath, 'audio.mp3', () => {
                 fs.unlinkSync(inputPath);
                 fs.unlinkSync(outputPath);
             });
         })
         .on('error', (err) => {
-            console.error(err);
+            console.error('FFMPEG ERROR:', err);
             res.status(500).send('Error al convertir');
         })
         .save(outputPath);
